# lang-portal/backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.Word:
    properties:
      created_at:
        type: string
      english:
        type: string
      hindi:
        type: string
      hinglish:
        type: string
      id:
        type: integer
      scrambled:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: support@languageportal.com
    name: Language Portal Support
  description: A backend API for managing language learning resources
  title: Language Portal API
  version: "1.0"
paths:
...
  /api/v1/words:
    get:
      description: Retrieve a paginated list of words with optional search and filtering
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: pageSize
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      - description: Filter by language
        in: query
        name: language
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Words retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List words
      tags:
      - words
    post:
      consumes:
      - application/json
      description: Create a new word with details like Hindi, English, Hinglish, etc.
      parameters:
      - description: Word details
        in: body
        name: word
        required: true
        schema:
          $ref: '#/definitions/models.Word'
      produces:
      - application/json
      responses:
        "201":
          description: Word created successfully
          schema:
            $ref: '#/definitions/models.Word'
        "400":
          description: Invalid request payload or validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new word
      tags:
      - words
  /api/v1/words/{id}:
    delete:
      description: Delete an existing word by its unique identifier
      parameters:
      - description: Word ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Word deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid word ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a word
      tags:
      - words
    get:
      description: Retrieve a word's details using its unique identifier
      parameters:
      - description: Word ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Word retrieved successfully
          schema:
            $ref: '#/definitions/models.Word'
        "400":
          description: Invalid word ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Word not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a word by ID
      tags:
      - words
    put:
      consumes:
      - application/json
      description: Update an existing word's details by its ID
      parameters:
      - description: Word ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated word details
        in: body
        name: word
        required: true
        schema:
          $ref: '#/definitions/models.Word'
      produces:
      - application/json
      responses:
        "200":
          description: Word updated successfully
          schema:
            $ref: '#/definitions/models.Word'
        "400":
          description: Invalid request payload or word ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a word
      tags:
      - words
  /api/v1/words/random:
    get:
      description: Retrieve a random word from the database
      produces:
      - application/json
      responses:
        "200":
          description: Random word retrieved successfully
          schema:
            $ref: '#/definitions/models.Word'
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a random word
      tags:
      - words
  /api/v1/words/search:
    get:
      description: Search for words using a query term and optional language filter
      parameters:
      - description: Search query
        in: query
        name: query
        required: true
        type: string
      - description: Language filter
        in: query
        name: language
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Search results retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search words
      tags:
      - words
...
schemes:
- http
swagger: "2.0"